feat(config): adding config / systemctl command - ability to change prefix (merge #42)
Many users requested to have the ability to change the prefix of the bot to a custom prefix. With using Keyv and SQLite, this issue has been solved and the feature is now added to the bot. It will be available after the 1.7 update.

BREAKING CHANGE: Custom Prefix ability

* added ability to change prefixes
fix #28 

* Changed PREFIX from config to DEFAULTPREFIX

* Updated all commands

* Tweaks and further improvements for finalizing the PR

Co-authored-by: Julian Yaman <julianyaman@posteo.eu>